requirements.txt
pyyaml>=5.3.1
setuptools>=49.2.0
awesome-pattern-matching>=0.19.0